requirements.txt
flask
python-dotenv
requests
pandas
pytest
requests-mock
flask-cors
groq
gunicorn
